app/requirements.txt
python-dotenv==1.0.1
openai==1.34.0
langchain==0.2.4
langchain-openai==0.1.8
fastapi==0.110.2
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic-settings==2.2.1
aiofiles==23.2.1
PyMuPDF==1.24.3
Motor==3.4.0
pydantic-mongo==2.3.0
pymongo[srv]
pydantic[email]
passlib==1.7.4
PyJWT==2.8.0
python-multipart==0.0.9
bcrypt
pytube 
langdetect
speechrecognition 
moviepy 
pydub
faiss-cpu
pinecone-client
langchain-community
openai-whisper
torch







